requirements.txt
Flask==3.1.3
flask_wtf==1.2.2
python-dotenv==1.2.1
WTForms==3.2.1
